+++ plugin/evm/network_handler.go
@@ -48,6 +48,10 @@ func (n networkHandler) HandleStateTrieLeafsRequest(ctx context.Context, nodeID 
 	return n.stateTrieLeafsRequestHandler.OnLeafsRequest(ctx, nodeID, requestID, leafsRequest)
 }
 
+func (n networkHandler) HandleLeafsRequest(ctx context.Context, nodeID ids.NodeID, requestID uint32, leafsRequest message.LeafsRequest) ([]byte, error) {
+	return n.stateTrieLeafsRequestHandler.OnLeafsRequest(ctx, nodeID, requestID, leafsRequest)
+}
+
 func (n networkHandler) HandleBlockRequest(ctx context.Context, nodeID ids.NodeID, requestID uint32, blockRequest message.BlockRequest) ([]byte, error) {
 	return n.blockRequestHandler.OnBlockRequest(ctx, nodeID, requestID, blockRequest)
 }